requirements.txt
opencv-python>=4.13.0
numpy>=2.4.2
onnxruntime>=1.23.2
mediapipe>=0.10.14
paho-mqtt>=2.1.0